requirements.txt
Flask~=3.0.2
Werkzeug~=3.0.1
psycopg2-binary
flask_sqlalchemy
flask_cors
rich~=13.7.1
pytz~=2024.1
pyyaml~=6.0.1
sqlalchemy~=2.0.28
python-dotenv~=1.0.1
waitress
bcrypt~=4.1.2
pyjwt~=2.8.0
flask_jwt_extended
waitress
pyopenssl
pydantic~=2.8.2
psycopg2~=2.9.9